backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pillow==10.1.0
qrcode[pil]==7.4.2
httpx==0.25.2
email-validator==2.1.0

# Face recognition dependencies
face-recognition==1.3.0
pgvector==0.2.4
numpy==1.24.3
opencv-python==4.8.1.78

# AWS dependencies
boto3==1.34.0
botocore==1.34.0
